backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
aiosqlite==0.20.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.1.1
pytest-asyncio==0.23.6
